+++ src/eval-frame/code-completion-request-response.js
@@ -1,4 +1,31 @@
-import { getCompletionScope } from "./parse-js-completion-scope";
+const scopeEndRE = /(\["[^"]*|\['[^']*|\[`[^`]*|\s*\.|\s*\.[_$a-zA-Z][_$a-zA-Z0-9]*)$/;
+const scopeRootRE = /[_$a-zA-Z][_$a-zA-Z0-9]*$/;
+const scopeMiddleRE = /\s*\.([_$a-zA-Z][_$a-zA-Z0-9]*)$|\["(.*)"\]$|\['(.*)'\]$|\[`(.*)`\]$|\[([0-9]*)\]$/;
+
+function getCompletionScope(code) {
+  let codeRemaining = code;
+  const scope = [];
+  let match = scopeEndRE.exec(code);
+  if (!match) return scope;
+
+  codeRemaining = codeRemaining.slice(0, match.index);
+  match = scopeMiddleRE.exec(codeRemaining);
+
+  let scopeItem;
+  while (match) {
+    // eslint-disable-next-line prefer-destructuring
+    scopeItem = match.slice(1).filter(x => x !== undefined)[0];
+
+    scope.unshift(scopeItem);
+    codeRemaining = codeRemaining.slice(0, match.index);
+    match = scopeMiddleRE.exec(codeRemaining);
+  }
+
+  const scopeRoot = scopeRootRE.exec(codeRemaining);
+  if (scopeRoot) scope.unshift(scopeRoot[0]);
+
+  return scope;
+}
 
 function getObjAtPath(path) {
   let obj = window;
@@ -8,16 +35,14 @@ function getObjAtPath(path) {
   return obj;
 }
 
-export function autocompleteJs(code) {
+function autocompleteJs(code) {
   const path = getCompletionScope(code);
   const obj = getObjAtPath(path);
   const items = Object.getOwnPropertyNames(obj);
-  console.log({ obj, path, items });
   return items;
 }
 
 export function codeCompletionRequestResponse(message) {
-  console.log(message);
   const { code, language } = message;
   if (language.languageId === "js") {
     return autocompleteJs(code);
